requirements.txt
mem0ai
google-genai
opencv-python
pyaudio
pillow
mss
python-dotenv